float/int conversions fixed

+++ src/InputHandler.cpp
@@ -75,13 +75,13 @@ void InputHandler::handleJoystickRemoved(const SDL_Event event) {
 
 void InputHandler::setJoystickValue(const int value, Vector2D* axisVector, Vector2DCoord coord) {
 	if (value > M_JOYSTICK_DEADZONE) {
-		axisVector->set(coord, 1);
+		axisVector->set(coord, 1.0);
 	}
 	else if (value < -M_JOYSTICK_DEADZONE) {
-		axisVector->set(coord, -1);
+		axisVector->set(coord, -1.0);
 	}
 	else {
-		axisVector->set(coord, 0);
+		axisVector->set(coord, 0.0);
 	}
 }
 
@@ -114,7 +114,7 @@ void InputHandler::initialiseJoystick(const int indexJoystick) {
 
 void InputHandler::clean() {
 	if (m_bJoysticksInitialised) {
-		int nbJoysticks = m_joysticks.size();
+		int nbJoysticks = (int) m_joysticks.size();
 		for (int i = 0; i < nbJoysticks; i++){
 			SDL_JoystickClose(m_joysticks[i].second);
 			m_joysticks[i].second = NULL;
@@ -141,10 +141,10 @@ void InputHandler::free() {
 int InputHandler::stickXValue(const int joyIndex, const JoystickControl stick) {
 	if (m_joystickAxisValues.size() > 0) {
 		if (stick == LEFT_STICK) {
-			return m_joystickAxisValues[m_joysticks[joyIndex].first].first.getX();
+			return (int) m_joystickAxisValues[m_joysticks[joyIndex].first].first.getX();
 		}
 		else if (stick == RIGHT_STICK) {
-			return m_joystickAxisValues[m_joysticks[joyIndex].first].second.getX();
+			return (int) m_joystickAxisValues[m_joysticks[joyIndex].first].second.getX();
 		}
 	}
 	return 0;
@@ -153,10 +153,10 @@ int InputHandler::stickXValue(const int joyIndex, const JoystickControl stick) {
 int InputHandler::stickYValue(const int joyIndex, const JoystickControl stick) {
 	if (m_joystickAxisValues.size() > 0) {
 		if (stick == LEFT_STICK) {
-			return m_joystickAxisValues[m_joysticks[joyIndex].first].first.getY();
+			return (int) m_joystickAxisValues[m_joysticks[joyIndex].first].first.getY();
 		}
 		else if (stick == RIGHT_STICK) {
-			return m_joystickAxisValues[m_joysticks[joyIndex].first].second.getY();
+			return (int) m_joystickAxisValues[m_joysticks[joyIndex].first].second.getY();
 		}
 	}
 	return 0;
